Seperates slice what has all about status with postcardSlice
Added postcardSlice and postcardSlice.test.js
